+++ modules/tinymce/src/plugins/emoticons/test/ts/browser/EmojiSearchTest.ts
@@ -1,6 +1,7 @@
 import { FocusTools, Keys, UiFinder, Waiter } from '@ephox/agar';
-import { describe, it } from '@ephox/bedrock-client';
-import { Attribute, SugarBody, SugarDocument, SugarElement } from '@ephox/sugar';
+import { before, describe, it } from '@ephox/bedrock-client';
+import { PlatformDetection } from '@ephox/sand';
+import { Attribute, SugarBody, SugarDocument } from '@ephox/sugar';
 import { TinyAssertions, TinyHooks, TinyUiActions } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
@@ -9,31 +10,37 @@ import Plugin from 'tinymce/plugins/emoticons/Plugin';
 
 import { fakeEvent } from '../module/test/Utils';
 
-describe('browser.tinymce.plugins.emoticons.ImageEmoticonTest', () => {
+describe('browser.tinymce.plugins.emoticons.EmojiSearchTest', () => {
+  before(function () {
+    // TODO: TINY-6905: Test is flaking on Chromium Edge 86, so we need to investigate
+    const platform = PlatformDetection.detect();
+    if (platform.browser.isChromium() && platform.os.isWindows()) {
+      this.skip();
+    }
+  });
+
   const hook = TinyHooks.bddSetupLight<Editor>({
     plugins: 'emoticons',
     toolbar: 'emoticons',
     base_url: '/project/tinymce/js/tinymce',
-    emoticons_database_url: '/project/tinymce/src/plugins/emoticons/main/js/emojiimages.js'
+    emoticons_database_url: '/project/tinymce/src/plugins/emoticons/main/js/emojis.js'
   }, [ Plugin ], true);
 
-  it('TBA: Open dialog, Search for "dog", Dog should be first option', async () => {
+  it('TBA: Open dialog, Search for "rainbow", Rainbow should be first option', async () => {
     const editor = hook.editor();
     const doc = SugarDocument.getDocument();
 
     TinyUiActions.clickOnToolbar(editor, 'button');
     await TinyUiActions.pWaitForDialog(editor);
     await FocusTools.pTryOnSelector('Focus should start on input', doc, 'input');
-    const input = FocusTools.setActiveValue(doc, 'dog');
+    const input = FocusTools.setActiveValue(doc, 'rainbow');
     fakeEvent(input, 'input');
     await Waiter.pTryUntil(
-      'Wait until dog is the first choice (search should filter)',
+      'Wait until rainbow is the first choice (search should filter)',
       () => {
         const item = UiFinder.findIn(SugarBody.body(), '.tox-collection__item:first').getOrDie();
-        const attr = Attribute.get(item, 'data-collection-item-value');
-        const img = SugarElement.fromHtml<HTMLImageElement>(attr);
-        const src = Attribute.get(img, 'src');
-        assert.equal(src, 'https://twemoji.maxcdn.com/v/13.0.1/72x72/1f436.png', 'Search should show a dog');
+        const value = Attribute.get(item, 'data-collection-item-value');
+        assert.equal(value, '🌈', 'Search should show rainbow');
       }
     );
     TinyUiActions.keydown(editor, Keys.tab());
@@ -41,13 +48,7 @@ describe('browser.tinymce.plugins.emoticons.ImageEmoticonTest', () => {
     TinyUiActions.keydown(editor, Keys.enter());
     await Waiter.pTryUntil(
       'Waiting for content update',
-      () => TinyAssertions.assertContentPresence(editor, {
-        'img[data-emoticon]': 1,
-        'img[data-mce-resize="false"]': 1,
-        'img[data-mce-placeholder="1"]': 1,
-        'img[alt="\ud83d\udc36"]': 1,
-        'img[src="https://twemoji.maxcdn.com/v/13.0.1/72x72/1f436.png"]': 1
-      })
+      () => TinyAssertions.assertContent(editor, '<p>🌈</p>')
     );
   });
 });